MerklizedSeq hierarchy moved into a dedicated file
